+++ src/ShellCheck/Analyzer.hs
@@ -20,10 +20,16 @@
 module ShellCheck.Analyzer (analyzeScript, ShellCheck.Analyzer.optionalChecks) where
 
 import ShellCheck.Analytics
+import ShellCheck.AST
+import ShellCheck.ASTLib (getPath)
 import ShellCheck.AnalyzerLib
 import ShellCheck.Interface
+import Control.Monad.Writer
 import Data.List
+import Data.Maybe
 import Data.Monoid
+import qualified Data.List.NonEmpty as NE
+import qualified Data.Map as Map
 import qualified ShellCheck.Checks.Commands
 import qualified ShellCheck.Checks.ControlFlow
 import qualified ShellCheck.Checks.Custom
@@ -33,12 +39,78 @@ import qualified ShellCheck.Checks.ShellSupport
 -- TODO: Clean up the cruft this is layered on
 analyzeScript :: AnalysisSpec -> AnalysisResult
 analyzeScript spec = newAnalysisResult {
-    arComments =
-        filterByAnnotation spec params . nub $
-            runChecker params (checkers spec params)
+    arComments = filterByAnnotation spec params rawComments ++ unusedComments
 }
   where
     params = makeParameters spec
+    rawComments = nub $ runChecker params (checkers spec params)
+    unusedCheckEnabled =
+        any (`elem` ["all", "check-unused-suppressions"]) (asOptionalChecks spec)
+    unsuppressedSpec = spec { asScript = stripDisableDirectives $ asScript spec }
+    unsuppressedParams = makeParameters unsuppressedSpec
+    suppressionCandidates =
+        if unusedCheckEnabled
+        then nub $ runChecker unsuppressedParams (checkers unsuppressedSpec unsuppressedParams)
+        else []
+    usedDirectives =
+        nub $ asUsedDisableDirectives spec ++ mapMaybe getSuppressingDirective suppressionCandidates
+    unusedComments =
+        if unusedCheckEnabled
+        then mapMaybe unusedComment $ getDisableDirectives (asScript spec)
+        else []
+
+    stripDisableDirectives = doTransform strip
+      where
+        strip (T_Annotation id annotations body) =
+            T_Annotation id (filter (not . isDisable) annotations) body
+        strip token = token
+        isDisable DisableComment {} = True
+        isDisable DisableCommentWithId {} = True
+        isDisable _ = False
+
+    getSuppressingDirective note = findSuppressor $ NE.toList path
+      where
+        path = getPath (parentMap params) (T_Bang $ tcId note)
+        code = cCode $ tcComment note
+        findSuppressor [] = Nothing
+        findSuppressor (T_Include {} : _) | not (asCheckSourced spec) = Nothing
+        findSuppressor (T_Annotation _ annotations _ : rest) =
+            case mapMaybe (suppressor code) annotations of
+                sourceId:_ -> sourceId
+                [] -> findSuppressor rest
+        findSuppressor (_:rest) = findSuppressor rest
+        suppressor code (DisableComment from to)
+            | code >= from && code < to = Just Nothing
+        suppressor code (DisableCommentWithId sourceId from to)
+            | code >= from && code < to = Just $ Just sourceId
+        suppressor _ _ = Nothing
+
+    getDisableDirectives :: Token -> [(Id, Integer, Integer)]
+    getDisableDirectives root = execWriter $ doAnalysis collect root
+      where
+        collect :: Token -> Writer [(Id, Integer, Integer)] ()
+        collect (T_Annotation _ annotations _) =
+            tell [(id, from, to) | DisableCommentWithId id from to <- annotations]
+        collect _ = return ()
+
+    unusedComment (id, from, to)
+        | not (shouldCheckDirective id) = Nothing
+        | id `elem` usedDirectives = Nothing
+        | otherwise = Just $ makeComment StyleC id 2337 (message from to)
+    shouldCheckDirective id
+        | asCheckSourced spec = True
+        | otherwise =
+            case (lookupFilename id, lookupFilename $ getId $ asScript spec) of
+                (Just file, Just rootFile) -> file == rootFile
+                _ -> True
+    lookupFilename id = posFile . fst <$> Map.lookup id (asTokenPositions spec)
+    message 0 1000000 =
+        "This disable directive is unnecessary because it suppresses no diagnostics here."
+    message from to
+        | to == from + 1 =
+            "This disable directive is unnecessary because SC" ++ show from ++ " is not reported here."
+        | otherwise =
+            "This disable directive is unnecessary because no diagnostics in its range are reported here."
 
 checkers spec params = mconcat $ map ($ params) [
     ShellCheck.Analytics.checker spec,
@@ -48,8 +120,15 @@ checkers spec params = mconcat $ map ($ params) [
     ShellCheck.Checks.ShellSupport.checker
     ]
 
-optionalChecks = mconcat $ [
-    ShellCheck.Analytics.optionalChecks,
-    ShellCheck.Checks.Commands.optionalChecks,
-    ShellCheck.Checks.ControlFlow.optionalChecks
+optionalChecks = unusedSuppressionCheck : mconcat [
+        ShellCheck.Analytics.optionalChecks,
+        ShellCheck.Checks.Commands.optionalChecks,
+        ShellCheck.Checks.ControlFlow.optionalChecks
     ]
+  where
+    unusedSuppressionCheck = newCheckDescription {
+        cdName = "check-unused-suppressions",
+        cdDescription = "Suggest removing disable directives that suppress no diagnostics",
+        cdPositive = "# shellcheck disable=SC2086\necho \"$var\"",
+        cdNegative = "# shellcheck disable=SC2086\necho $var"
+    }